DAY_3/LAB_4/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-core>=0.3.0
chromadb>=0.5.0
openai>=1.40.0
python-dotenv>=1.0.0
pydantic>=2.8.0
